Rename infoutil.GetInfo() to info.New()
